% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_new_locations_for_spatpca_object}
\alias{check_new_locations_for_spatpca_object}
\title{Internal function: Validate new locations for a spatpca object}
\usage{
check_new_locations_for_spatpca_object(spatpca_object, x_new)
}
\arguments{
\item{spatpca_object}{An \code{spatpca} class object}

\item{x_new}{New location matrix.}
}
\description{
Internal function: Validate new locations for a spatpca object
}
\keyword{internal}
