% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{checkInputData}
\alias{checkInputData}
\title{Internal function: Validate input data for a spatpca object}
\usage{
checkInputData(Y, x, M)
}
\arguments{
\item{Y}{Data matrix}

\item{x}{Location matrix.}

\item{M}{Number of folds for cross-validation}
}
\value{
\code{NULL}.
}
\description{
Internal function: Validate input data for a spatpca object
}
\keyword{internal}
