% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{checkNewLocationsForSpatpcaObject}
\alias{checkNewLocationsForSpatpcaObject}
\title{Internal function: Validate new locations for a spatpca object}
\usage{
checkNewLocationsForSpatpcaObject(spatpca_object, x_new)
}
\arguments{
\item{spatpca_object}{An \code{spatpca} class object}

\item{x_new}{New location matrix.}
}
\description{
Internal function: Validate new locations for a spatpca object
}
\keyword{internal}
