% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{detrend}
\alias{detrend}
\title{Internal function: Detrend Y by column-wise centering}
\usage{
detrend(Y, is_Y_detrended)
}
\arguments{
\item{Y}{Data matrix}
}
\value{
Detrended data matrix
}
\description{
Internal function: Detrend Y by column-wise centering
}
\keyword{internal}
