% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenEstimate}
\alias{eigenEstimate}
\title{Internal function: Estimate eigenfunctions}
\usage{
eigenEstimate(phir, Yr, gamma, phi2r)
}
\arguments{
\item{phir}{A location matrix for original observations}

\item{Yr}{A data matrix}

\item{gammar}{A gamma value}

\item{phir2}{A location matrix used to predict}
}
\value{
A eignfunction estimate matrix
}
\description{
Internal function: Estimate eigenfunctions
}
\keyword{internal}
