% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fetchUpperBoundNumberEigenfunctions}
\alias{fetchUpperBoundNumberEigenfunctions}
\title{Internal function: Fetch the upper bound of the number of eigenfunctions}
\usage{
fetchUpperBoundNumberEigenfunctions(Y, M)
}
\arguments{
\item{Y}{Data matrix}

\item{M}{Number of folds for cross-validation}
}
\value{
integer
}
\description{
Internal function: Fetch the upper bound of the number of eigenfunctions
}
\keyword{internal}
