% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{scaleLocation}
\alias{scaleLocation}
\title{Internal function: Scale one-dimension locations}
\usage{
scaleLocation(location)
}
\arguments{
\item{location}{Location matrix}
}
\value{
scaled location matrix
}
\description{
Internal function: Scale one-dimension locations
}
\keyword{internal}
