% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{set_cores}
\alias{set_cores}
\title{Internal function: Set the number of cores for parallel computing}
\usage{
set_cores(ncores = NULL)
}
\arguments{
\item{maxit}{Number of number of cores for parallel computing.}
}
\value{
Logical
}
\description{
Internal function: Set the number of cores for parallel computing
}
\keyword{internal}
