% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{setCores}
\alias{setCores}
\title{Internal function: Set the number of cores for parallel computing}
\usage{
setCores(ncores = NULL)
}
\arguments{
\item{ncores}{Number of number of cores for parallel computing. Default is NULL.}
}
\value{
Logical
}
\description{
Internal function: Set the number of cores for parallel computing
}
\keyword{internal}
