% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{setGamma}
\alias{setGamma}
\title{Internal function: Set tuning parameter - gamma}
\usage{
setGamma(gamma, Y)
}
\arguments{
\item{gamma}{Vector of a nonnegative hyper parameter sequence for tuning eigenvalues. Default is NULL.}

\item{Y}{Data matrix}
}
\value{
Modified vector of a nonnegative hyper parameter sequence for tuning eigenvalues.
}
\description{
Internal function: Set tuning parameter - gamma
}
\keyword{internal}
