% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{setL2}
\alias{setL2}
\title{Internal function: Set tuning parameter - l2}
\usage{
setL2(tau2)
}
\arguments{
\item{tau2}{numeric vector of a nonnegative sparseness parameter sequence. Default is NULL.}
}
\value{
modified numeric vector of a nonnegative tuning parameter sequence for ADMM use
}
\description{
Internal function: Set tuning parameter - l2
}
\keyword{internal}
