% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{setNumberEigenfunctions}
\alias{setNumberEigenfunctions}
\title{Internal function: Set the number of eigenfunctions for a spatpca object}
\usage{
setNumberEigenfunctions(K, Y, M)
}
\arguments{
\item{K}{Optional user-supplied number of eigenfunctions.}

\item{Y}{Data matrix}

\item{M}{Number of folds for cross-validation}
}
\value{
integer
}
\description{
Internal function: Set the number of eigenfunctions for a spatpca object
}
\keyword{internal}
