% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{setTau1}
\alias{setTau1}
\title{Internal function: Set tuning parameter - tau1}
\usage{
setTau1(tau1, M)
}
\arguments{
\item{tau1}{Vector of a nonnegative smoothness parameter sequence. Default is NULL.}

\item{M}{Number of folds for cross-validation}
}
\value{
Modified vector of a nonnegative smoothness parameter sequence.
}
\description{
Internal function: Set tuning parameter - tau1
}
\keyword{internal}
