% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{setTau2}
\alias{setTau2}
\title{Internal function: Set tuning parameter - tau2}
\usage{
setTau2(tau2, M)
}
\arguments{
\item{tau2}{numeric vector of a nonnegative sparseness parameter sequence. Default is NULL.}

\item{M}{Number of folds for cross-validation}
}
\value{
modified numeric vector of a nonnegative sparseness parameter sequence.
}
\description{
Internal function: Set tuning parameter - tau2
}
\keyword{internal}
