% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spatialPrediction}
\alias{spatialPrediction}
\title{Internal function: Spatial prediction}
\usage{
spatialPrediction(phir, Yr, gamma, predicted_eignefunction)
}
\arguments{
\item{phir}{A matrix of estimated eigenfunctions based on original locations}

\item{Yr}{A data matrix}

\item{predicted_eignefunction}{A vector of values of an eigenfunction on new locations}

\item{gammar}{A gamma value}
}
\value{
A list of objects
\item{prediction}{A vector of spatial predicitons}
\item{estimated_covariance}{An estimated covariance matrix.}
\item{eigenvalue}{A vecotor of estimated eigenvalues.}
\item{error}{Error rate for the ADMM algorithm}
}
\description{
Internal function: Spatial prediction
}
\keyword{internal}
