\name{SpatPCA-package}
\alias{SpatPCA-package}
\docType{package}
\title{Regularized Principal Component Analysis for Spatial Data}
\description{
A new regularization approach to estimate the leading spatial patterns via smoothness and sparseness penalties. 
}
\details{
\tabular{ll}{
Package:\tab SpatPCA\cr
Type:\tab  Package\cr
Version:\tab   1.0.0.1\cr
Date:\tab   2015-02-05\cr
License: \tab GPL version 2 or newer\cr
}
}
\author{ 
Wen-Ting Wang <egpivo@gmail.com> and Hsin-Cheng Huang <hchuang@stat.sinica.edu.tw>
}