% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{thinPlateSplineMatrix}
\alias{thinPlateSplineMatrix}
\title{Thin-plane spline matrix}
\usage{
thinPlateSplineMatrix(location)
}
\arguments{
\item{location}{A location matrix}
}
\value{
A thin-plane spline matrix
}
\description{
Produce a thin-plane spline matrix based on a given location matric
}
\examples{
pesudo_sequence <- seq(-5, 5, length = 5)
two_dim_location <- as.matrix(expand.grid(x = pesudo_sequence, y = pesudo_sequence))
thin_plate_matrix <- thinPlateSplineMatrix(two_dim_location)
}
