% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tpm2}
\alias{tpm2}
\title{Internal function: predicted thin-plane spline matrix}
\usage{
tpm2(z, P, Phi)
}
\arguments{
\item{P}{A location matrix}

\item{Phi}{An eigenvector matrix}
}
\value{
A predicted thin-plane spline matrix
}
\description{
Internal function: predicted thin-plane spline matrix
}
\keyword{internal}
