% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tpmatrix}
\alias{tpmatrix}
\title{Internal function: thin-plane spline matrix}
\usage{
tpmatrix(P)
}
\arguments{
\item{P}{A location matrix}
}
\value{
A thin-plane spline matrix
}
\description{
Internal function: thin-plane spline matrix
}
\keyword{internal}
