% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPCA-package.R
\docType{package}
\name{SpatPCA-package}
\alias{SpatPCA-package}
\alias{SpatPCA}
\title{SpatPCA: Regularized Principal Component Analysis for Spatial Data}
\description{
Provides regularized principal component analysis for spatial datasets with smoothness, sparseness, and orthogonality constraints handled via an ADMM algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/egpivo/SpatPCA}
  \item Report bugs at \url{https://github.com/egpivo/SpatPCA/issues}
}

}
\author{
\strong{Maintainer}: Wen-Ting Wang \email{egpivo@gmail.com} (\href{https://orcid.org/0000-0003-3051-7302}{ORCID})

Authors:
\itemize{
  \item Hsin-Cheng Huang \email{hchuang@stat.sinica.edu.tw} (\href{https://orcid.org/0000-0002-5613-349X}{ORCID})
}

}
