\name{SpatPCA-package}
\alias{SpatPCA-package}
\docType{package}
\title{Regularized Principal Component Analysis for Spatial Data}
\description{
A new regularization approach to estimate the leading spatial patterns via smoothness and sparseness penalties, and spatial predctions for spatial data that may be irregularly located in space (including 1D, 2D and 3D), and obtain the spatial prediction at the designated locations.
}
\details{\tabular{ll}{
Package:\tab SpatPCA\cr
Type:\tab Package\cr
Version:\tab  1.3.1.0\cr
Date:\tab 2021-01-19\cr
License: \tab GPL version 3\cr
}
}
\author{ 
Wen-Ting Wang <egpivo@gmail.com> and Hsin-Cheng Huang <hchuang@stat.sinica.edu.tw>
}