% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatPCA.R
\name{spatpca}
\alias{spatpca}
\title{Regularized PCA for spatial data}
\usage{
spatpca(
  x,
  Y,
  M = 5,
  K = NULL,
  K.select = ifelse(is.null(K), TRUE, FALSE),
  tau1 = NULL,
  tau2 = NULL,
  gamma = NULL,
  x_new = NULL,
  center = FALSE,
  plot.cv = FALSE,
  maxit = 100,
  thr = 1e-04
)
}
\description{
Regularized PCA for spatial data
}
