% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spatpcaCV}
\alias{spatpcaCV}
\title{Internal function: M-fold Cross-validation}
\usage{
spatpcaCV(sxyr, Yr, M, K, tau1r, tau2r, gammar, nkr, maxit, tol, l2r)
}
\arguments{
\item{sxyr}{A location matrix}

\item{Yr}{A data matrix}

\item{M}{The number of folds for CV}

\item{K}{The number of estimated eigen-functions}

\item{tau1r}{A range of tau1}

\item{tau2r}{A range of tau2}

\item{gammar}{A range of gamma}

\item{nkr}{A vector of fold numbers}

\item{maxit}{A maximum number of iteration}

\item{tol}{A tolerance rate}

\item{l2r}{A given tau2}
}
\value{
A list of selected parameters
}
\description{
Internal function: M-fold Cross-validation
}
\keyword{internal}
